server fixed

+++ lib/Chat/Chat.php
@@ -5,6 +5,11 @@ import('core.File');
 module('model.ChatMessage');
 class Chat extends Flow
 {
+    private static $static_flag = false;
+    
+    public static function __shutdown__(){
+        self::send_sig();
+    }
     public function models(){
         $paginator = new Paginator(50, $this->inVars('page', 1));
         $object_list = C(ChatMessage)->find_all(Q::order('-id'), $paginator);
@@ -20,11 +25,14 @@ class Chat extends Flow
             $message->set_model($this->vars());
             $message->save();
             C($message)->commit();
-            self::send_sig();
+            self::$static_flag = true;
         }
         exit;
     }
     private static function send_sig(){
+        if(self::$static_flag == false) {
+            return ;
+        }
         $pids = self::get_pids();
         foreach($pids as $pid){
             posix_kill($pid, SIGUSR1);
